app/requirements.txt
alembic==1.13.1
blessed==1.20.0
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
editor==1.6.6
greenlet==3.0.3
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
inquirer==3.2.4
install==1.3.5
Mako==1.3.5
MarkupSafe==2.1.5
migrate==0.3.8
mysql-connector-python==8.4.0
mysqlclient==2.2.4
pbr==6.0.0
python-dotenv==1.0.1
readchar==4.1.0
requests==2.31.0
runs==1.2.2
six==1.16.0
SQLAlchemy==2.0.30
sqlalchemy-migrate==0.13.0
sqlparse==0.5.0
Tempita==0.5.2
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
xmod==1.8.1
zipp==3.18.2
